modulized primitive editor
