+++ frontend/src/app/modules/canvasEditor/boxEditors/hooks/useState.tsx
@@ -1,0 +1,21 @@
+import { PrimitiveType } from "../../shared/types";
+import { useState } from "react";
+
+export const useGlobalStates = () => {
+  const [hoverRemove, setHoverRemove] = useState(false);
+  return {
+    hoverRemove,
+    setHoverRemove,
+  };
+};
+
+export const usePrimitiveStates = (element: any) => {
+  const [dataType, setDataType] = useState<PrimitiveType>(element.kind.type);
+  const [contentValue, setContentValue] = useState(element.kind.value);
+  return {
+    dataType,
+    setDataType,
+    contentValue,
+    setContentValue,
+  };
+};
