+++ frontend/src/app/modules/canvasEditor/boxEditors/utils/BoxEditorConfigs.tsx
@@ -1,0 +1,3 @@
+export const BoxEditorConfigs = {
+  primitive: {},
+};
