+++ frontend/src/app/modules/canvasEditor/shared/types.ts
@@ -1,6 +1,13 @@
 export type PrimitiveType = "None" | "int" | "float" | "str" | "bool";
 export type CollectionType = "list" | "tuple" | "set" | "dict";
 export type SpecialType = "function";
+export type BoxType =
+  | PrimitiveKind
+  | FunctionKind
+  | ListKind
+  | TupleKind
+  | SetKind
+  | DictKind;
 
 export type ValueType = PrimitiveType | CollectionType | SpecialType;
 
@@ -42,17 +49,15 @@ export interface DictKind {
   value: Record<number, number | null>;
 }
 
-export type ElementKind =
-  | PrimitiveKind
-  | FunctionKind
-  | ListKind
-  | TupleKind
-  | SetKind
-  | DictKind;
-
 export interface CanvasElement {
   id: number | string;
   x: number;
   y: number;
-  kind: ElementKind;
+  kind: BoxType;
+}
+
+export interface BoxEditorType {
+  metadata: { id: string; kind: PrimitiveKind };
+  onSave: (kind: PrimitiveKind) => void;
+  onRemove: () => void;
 }
